(fix) error in MidUps list page with scope issue 1. fixed deleted midups not being update properly in the DOM
