+++ client/index.html
@@ -115,6 +115,7 @@
           <script src="app/admin/admin.js"></script>
           <script src="app/google-maps/google-maps.controller.js"></script>
           <script src="app/google-maps/google-maps.js"></script>
+          <script src="app/google-maps/midup.controller.js"></script>
           <script src="app/google-maps/scripts/v3_epoly.js"></script>
           <script src="app/landing/js/bootstrap.js"></script>
           <script src="app/landing/js/countUp.min.js"></script>
@@ -131,7 +132,6 @@
           <script src="app/landing/js/zalive.scripts.js"></script>
           <script src="app/main/main.controller.js"></script>
           <script src="app/main/main.js"></script>
-          <script src="app/rooms/midup.controller.js"></script>
           <script src="app/rooms/rooms-routes.js"></script>
           <script src="app/rooms/rooms.controller.js"></script>
           <script src="app/services/services.js"></script>
